AJout q3 model 1 et 3 ne marche comme voulu valeur fausse)

+++ prog/tri_fusion.c
@@ -4,24 +4,24 @@
 #include <math.h>
 #include <unistd.h>
 
-void fusion(long double tableau[],int deb1,int fin1,int fin2)
+void fusion(double tableau[],int deb1,int fin1,int fin2)
         {
-        long double *table1;
+        double *table1;
         int deb2=fin1+1;
         int compt1=deb1;
         int compt2=deb2;
         int i;
-        
-        table1=malloc((fin1-deb1+1)*sizeof(long double));
+
+        table1=malloc((fin1-deb1+1)*sizeof(double));
 
         //on recopie les éléments du début du tableau
         for(i=deb1;i<=fin1;i++)
             {
             table1[i-deb1]=tableau[i];
             }
-                        
+
         for(i=deb1;i<=fin2;i++)
-            {        
+            {
             if (compt1==deb2) //c'est que tous les éléments du premier tableau ont été utilisés
                 {
                 break; //tous les éléments ont donc été classés
@@ -44,7 +44,7 @@ void fusion(long double tableau[],int deb1,int fin1,int fin2)
             }
         free(table1);
         }
-void tri_fusion_bis(long double tableau[],int deb,int fin)
+void tri_fusion_bis(double tableau[],int deb,int fin)
         {
         if (deb!=fin)
             {
@@ -55,12 +55,10 @@ void tri_fusion_bis(long double tableau[],int deb,int fin)
             }
         }
 
-void tri_fusion(long double tableau[],int longueur)
+void tri_fusion(double tableau[],int longueur)
      {
      if (longueur>0)
             {
             tri_fusion_bis(tableau,0,longueur-1);
             }
      }
-   
-
